Refactor primitive parsing
This commit refactors the way primitives are parsed, in preparation for
the subsequent implementation of primitive aliases (RFC #44). Before
this commit, primitive declarations in both structures and signatures
are parsed into the same [value_description] record as regular OCaml
values. As such, both are represented by the [Psig_value] variant in
signatures. After this commit, primitive declarations are parsed into a
separate [primitive_description] record, and are represented by a
new [Psig_primitive] variant in signatures. In structures,
[Pstr_primitive] is  updated to contain a [primitive_description] rather
than a [value_description].

The only observable change from this commit should be that the form [val
v : t] is now rejected by the parser, rather than the typechecker. The
code path which parsed this form no longer makes any sense after this
change.

There are some additional changes to tests which print out
representations of the AST, and one test (docstrings.ml) which parsed
but did not typecheck the now-invalid [val v : t] form in a structure
has been updated to instead use the form [let v = e].