Add primitive aliases to unsafe rec mod check
@lpw25 pointed out that since we don't enforce any constraints on
primitive alias type ascriptions, one could use recursive modules to
alias a primitive with a different and potentially unsafe type.

We address this by checking if the aliased primitive's type is more
general than that of its alias, and if not, then we ferry this
information to [Translmod], where we can raise an error via the
recursive module safety codepath.

We only perform this check in signatures; in structures, we continue to
mandate that the types unify.